+++ src/microbe_stage/editor/CombinedEditorAction.cs
@@ -58,7 +58,7 @@ public class CombinedEditorAction : EditorAction
             action.UndoAction();
     }
 
-    public override int CalculateCost()
+    public override double CalculateCost()
     {
         return Actions.Sum(a => a.CalculateCost());
     }
